refactor HTMLTemplate to PrintEntry
